+++ insts/std/IIR/BUZZ.cpp
@@ -23,7 +23,7 @@ BUZZ::~BUZZ()
 }
 
 
-int BUZZ::init(float p[], short n_args)
+int BUZZ::init(float p[], int n_args)
 {
 // p0 = start; p1 = duration; p2 = amplitude; p3 = pitch (hz or oct.pc)
 // p4 = stereo spread (0-1) [optional]
